requirements.txt
gradio==4.28.3
torch
numpy
opencv-python-headless
setuptools
Pillow
torchvision
addict
future
lmdb
pyyaml
requests
scikit-image
scipy
tb-nightly
tqdm
yapf
psutil
ffmpeg-python
huggingface_hub